+++ src-tauri/Cargo.toml
@@ -25,6 +25,7 @@ anyhow = "1.0.89"
 windows = {version = "0.58.0", features = ["Win32", "Win32_System", "Win32_System_Wmi", "Win32_System_Com", "Win32_Security"] }
 wmi = "0.14.4"
 lazy_static = "1.5.0"
+tauri-plugin-dialog = "2"
 [features]
 custom-protocol = ["tauri/custom-protocol"]
