+++ src-tauri/NUCtoolChange.md
@@ -1,188 +1,139 @@
 # NUCtool（Changelog）
 
-
+## [0.3.5-<font color=red>Alpha</font>] - 2025-1-30
+#### 新特性
+- 询问是否更新
+- 代码优化
+#### 依赖更新
+- 更新最新库
 
 ## [0.3.4-<font color=red>Alpha</font>] - 2025-1-26
-
-### 修复
+#### 修复
 - LAPKC71F 获取转速异常
 
+## [0.3.4-<font color=red>Alpha</font>] - 2025-1-26
+#### 修复
+- LAPKC71F 获取转速异常
 
-
-## [0.3.3-<font color=green></font>] - 2025-1-24
-
-### 测试版本号
-
-
+## [0.3.3-<font color=pink>Beta</font>] - 2025-1-24
+#### 测试版本号
 
 ## [0.3.2-<font color=red>Alpha</font>] - 2025-1-24
-
-### 新特性
-- [x] 托盘菜单增加调试选项
-- [x] 自动检测更新(Beta)
-
+#### 新特性
+- 托盘菜单增加调试选项
+- 自动检测更新(Beta)
 
 ## [0.3.1-<font color=red>Alpha</font>] - 2025-1-22
-
-### 新特性
-- [x] 适配LAPKC71F风扇测试
-- [x] 自动编译测试
-
-
+#### 新特性
+- 适配LAPKC71F风扇测试
+- 自动编译测试
 
 ## [0.3.0-<font color=red>Alpha</font>] - 2024-10-27
-
-### 新特性
+#### 新特性
 - 新增TDP工具(经过LAPAC71H机型测试)
   - `加载配置` 后 设置 `%AppData%\com.nuc.x15.fan.cyear.app\debug.config` 值为 `1` 开启
-
-### 修复
+#### 修复
 - 禁用自动加载配置(更改变量导致旧配置失效)
-
-### 性能改进
+#### 性能改进
 - 主入口代码模块化
 
-
-
 ## [0.2.1-<font color=red>Alpha</font>] - 2024-10-14
-
-### 完美支持机型
-- [x] LAPAC71H
-- [ ] LAPKC71F(风扇相反？)
-    > - [x] 风扇状态检测适配
-    > - [x] 风扇异常处理
-    > - [ ] ？？？？
-    > - [ ] BUG 修复
-- [ ] 其余机型(风扇相反？数据未知？)
-
-### 新特性
+#### 完美支持机型
+- [X]  LAPAC71H
+- [ ]  LAPKC71F(风扇相反？)
+  > - [X]  风扇状态检测适配
+  > - [X]  风扇异常处理
+  > - [ ]  ？？？？
+  > - [ ]  BUG 修复
+  >
+- [ ]  其余机型(风扇相反？数据未知？)
+#### 新特性
 - 针对LAPAC71H机型优化(<font color=pink>Beta</font>)
-
-### 修复
+#### 修复
 - 异常修复
 - 逻辑错误
 - 代码错误
-
-### 性能改进
+#### 性能改进
 - 优化运行性能
 - 优化代码逻辑
 - 优化调度细节
-
-### <font color=red>已知BUG</font>
-- 退出程序风扇未恢复自动
-    ###### 概率问题(待修复)
+#### <font color=red>已知BUG</font>
+- 退出程序风扇未恢复自动###### 概率问题(待修复)
 - <font color=red>获取设备数据异常(疑似反复实例wmi导致，待解决)</font>
 
-
-
 ## [0.2.0-<font color=red>Alpha</font>] - 2024-10-12
-
-### 新特性
-- [x] 实时风扇转速图表(<font color=pink>Beta</font>)
-- [x] 实时设备温度图表(<font color=pink>Beta</font>)
-- [x] 曲线图表
-- [x] 风扇灵敏度`1`℃
-- [x] 绕过杀毒软件
-
-### 修复
+#### 新特性
+- [X]  实时风扇转速图表(<font color=pink>Beta</font>)
+- [X]  实时设备温度图表(<font color=pink>Beta</font>)
+- [X]  曲线图表
+- [X]  风扇灵敏度`1`℃
+- [X]  绕过杀毒软件
+#### 修复
 - 控制风扇转速代码
 - 获取风扇转速代码
 - 代码错误导致转速异常
-
-### 性能改进
+#### 性能改进
 - 优化运行性能
 - 优化代码逻辑
 - 优化界面细节
 - 优化调度细节
-
-### <font color=red>已知BUG</font>
+#### <font color=red>已知BUG</font>
 - 风扇问题
 > - 左风扇在温度过高过低速度未按照配置设定（右未知）
 > - 多次提示风扇状态异常
 - 托盘退出时风扇未恢复自动
 
-
-
 ## [0.1.5-<font color=red>Alpha</font>] - 2024-10-09
-
-### 新特性
+#### 新特性
 - 实时风扇转速图表(<font color=pink>Beta</font>)
-
-### 修复
+#### 修复
 - 修复了严重逻辑错误
-
-### 性能改进
+#### 性能改进
 - 优化运行性能
 - 优化代码逻辑
 - 优化界面
 
-
-
 ## [0.1.4-<font color=red>Alpha</font>] - 2024-10-09
-
-### 新特性
+#### 新特性
 - 自动提权
 - 风扇状态异常自动恢复
 
-
-
 ## [0.1.3-<font color=red>Alpha</font>] - 2024-10-09
-
-### 新特性
+#### 新特性
 - 使用最新版本界面库
 - 美化界面
 - 更改程序名称
 - 系统托盘
-
-
-
 ## [0.1.2-<font color=red>Alpha</font>] - 2024-10-08
-
-### 新特性
+#### 新特性
 - 合并所有代码
 - 支持风扇满转
 - 更改程序名称
-
-### 性能改进
+#### 性能改进
 - 优化性能
 
-
-
 ## [0.1.1-<font color=red>Alpha</font>] - 2024-10-07
-
-### 新特性
+#### 新特性
 - 资源打包
-
-### 修复
+#### 修复
 - 温度过高超过图表导致风扇不正常工作
 - 温度异常导致控制失效
 
-
-
 ## [0.1.0-<font color=pink>Beta</font>] - 2024-10-06
-
-### 新特性
+#### 新特性
 - 增加了风扇保存加载配置文件
-
-### 修复
+#### 修复
 - 修复了风扇停转致命问题
-
-### 性能改进
+#### 性能改进
 - 优化了使用性能
-
-### 依赖更新
+#### 依赖更新
 - 支持msi安装
-
-### <font color=red>已知BUG</font>
-
-- [ ] 温度过高超过图表导致风扇不正常工作
-- [ ] 合并所有代码，提升性能
-- [ ] 支持风扇满转
-
-
+#### <font color=red>已知BUG</font>
+- [ ]  温度过高超过图表导致风扇不正常工作
+- [ ]  合并所有代码，提升性能
+- [ ]  支持风扇满转
 
 ## 使用方法0.3.0版本适配({}中为必须步骤，[]可选不建议执行步骤)
-  - 打开程序后`调整曲线`并点击`{保存配置}`
-  - 每次重新打开程序需要点击`{加载配置}`
-  - [`加载配置` 后 设置 `%AppData%\com.nuc.x15.fan.cyear.app\debug.config` 值为 `1`] 开启 TDP 测试工具
-  ######  PS: 未设置TDP 首次读取全部为`0`是正常的，显卡只支持Arc GPU
+- 打开程序后`{调整曲线}`并点击`{保存配置}`
+- 每次重新打开程序需要点击`加载配置`
+- 点击`Start`
