+++ src-tauri/src/modules/mod.rs
@@ -1,4 +1,7 @@
 pub mod config;
+pub mod fan;
+pub mod permissions;
 pub mod setup;
 pub mod struct_set;
+pub mod tdp;
 pub mod wmi;
